feat : WebSocketUserTracker DIP화
