+++ src/main/java/kimp/websocket/service/WebSocketUserTracker.java
@@ -1,50 +1,38 @@
 package kimp.websocket.service;
 
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.context.event.EventListener;
-import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
-import org.springframework.stereotype.Service;
 import org.springframework.web.socket.messaging.SessionConnectEvent;
 import org.springframework.web.socket.messaging.SessionDisconnectEvent;
 
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.atomic.AtomicInteger;
-
-@Service
-@Slf4j
-public class WebSocketUserTracker {
+/**
+ * WebSocket 사용자 연결 추적 서비스 인터페이스
+ */
+public interface WebSocketUserTracker {
     
-    private final ConcurrentHashMap<String, String> sessionRegistry = new ConcurrentHashMap<>();
-    private final AtomicInteger userCount = new AtomicInteger(0);
+    /**
+     * WebSocket 연결 이벤트 처리
+     * 
+     * @param event 연결 이벤트
+     */
+    void handleWebSocketConnectListener(SessionConnectEvent event);
     
-    @EventListener
-    public void handleWebSocketConnectListener(SessionConnectEvent event) {
-        StompHeaderAccessor headerAccessor = StompHeaderAccessor.wrap(event.getMessage());
-        String sessionId = headerAccessor.getSessionId();
-        
-        if (sessionId != null && !sessionRegistry.containsKey(sessionId)) {
-            sessionRegistry.put(sessionId, sessionId);
-            int currentCount = userCount.incrementAndGet();
-            log.info("WebSocket 사용자 연결: {} (총 사용자 수: {})", sessionId, currentCount);
-        }
-    }
+    /**
+     * WebSocket 연결 해제 이벤트 처리
+     * 
+     * @param event 연결 해제 이벤트
+     */
+    void handleWebSocketDisconnectListener(SessionDisconnectEvent event);
     
-    @EventListener
-    public void handleWebSocketDisconnectListener(SessionDisconnectEvent event) {
-        StompHeaderAccessor headerAccessor = StompHeaderAccessor.wrap(event.getMessage());
-        String sessionId = headerAccessor.getSessionId();
-        
-        if (sessionId != null && sessionRegistry.remove(sessionId) != null) {
-            int currentCount = userCount.decrementAndGet();
-            log.info("WebSocket 사용자 연결 해제: {} (총 사용자 수: {})", sessionId, currentCount);
-        }
-    }
+    /**
+     * 현재 연결된 사용자 수 조회
+     * 
+     * @return 현재 연결된 사용자 수
+     */
+    int getUserCount();
     
-    public int getUserCount() {
-        return userCount.get();
-    }
-    
-    public int getActiveSessionCount() {
-        return sessionRegistry.size();
-    }
+    /**
+     * 현재 활성 세션 수 조회
+     * 
+     * @return 현재 활성 세션 수
+     */
+    int getActiveSessionCount();
 }
